requirements.txt
fastapi==0.111.0
uvicorn==0.25.0
psycopg[binary]==3.2.11
python-dotenv==1.0.0
